crimson/os/seastore: move lba/backref_map_val_t to seastore_types.h
