+++ src/crimson/os/seastore/CMakeLists.txt
@@ -12,7 +12,6 @@ set(crimson_seastore_srcs
   lba_manager.cc
   async_cleaner.cc
   backref_manager.cc
-  backref/backref_tree_node.cc
   backref/btree_backref_manager.cc
   lba_manager/btree/btree_lba_manager.cc
   lba_manager/btree/lba_btree_node.cc
