+++ src/crimson/os/seastore/lba_manager/btree/lba_btree_node.h
@@ -28,30 +28,6 @@ using LBANode = FixedKVNode<laddr_t>;
 
 class BtreeLBAMapping;
 
-/**
- * lba_map_val_t
- *
- * struct representing a single lba mapping
- */
-struct lba_map_val_t {
-  extent_len_t len = 0;  ///< length of mapping
-  pladdr_t pladdr;         ///< physical addr of mapping or
-			   //	laddr of a physical lba mapping(see btree_lba_manager.h)
-  extent_ref_count_t refcount = 0; ///< refcount
-  uint32_t checksum = 0; ///< checksum of original block written at paddr (TODO)
-
-  lba_map_val_t() = default;
-  lba_map_val_t(
-    extent_len_t len,
-    pladdr_t pladdr,
-    extent_ref_count_t refcount,
-    uint32_t checksum)
-    : len(len), pladdr(pladdr), refcount(refcount), checksum(checksum) {}
-  bool operator==(const lba_map_val_t&) const = default;
-};
-
-std::ostream& operator<<(std::ostream& out, const lba_map_val_t&);
-
 constexpr size_t LBA_BLOCK_SIZE = 4096;
 
 using lba_node_meta_t = fixed_kv_node_meta_t<laddr_t>;
@@ -117,30 +93,6 @@ using LBAInternalNodeRef = LBAInternalNode::Ref;
  * TODO: the above alignment probably isn't portable without further work
  */
 constexpr size_t LEAF_NODE_CAPACITY = 140;
-
-/**
- * lba_map_val_le_t
- *
- * On disk layout for lba_map_val_t.
- */
-struct __attribute__((packed)) lba_map_val_le_t {
-  extent_len_le_t len = init_extent_len_le(0);
-  pladdr_le_t pladdr;
-  extent_ref_count_le_t refcount{0};
-  ceph_le32 checksum{0};
-
-  lba_map_val_le_t() = default;
-  lba_map_val_le_t(const lba_map_val_le_t &) = default;
-  explicit lba_map_val_le_t(const lba_map_val_t &val)
-    : len(init_extent_len_le(val.len)),
-      pladdr(pladdr_le_t(val.pladdr)),
-      refcount(val.refcount),
-      checksum(val.checksum) {}
-
-  operator lba_map_val_t() const {
-    return lba_map_val_t{ len, pladdr, refcount, checksum };
-  }
-};
 
 struct LBALeafNode
   : FixedKVLeafNode<
@@ -308,7 +260,6 @@ using LBALeafNodeRef = TCachedExtentRef<LBALeafNode>;
 
 #if FMT_VERSION >= 90000
 template <> struct fmt::formatter<crimson::os::seastore::lba_manager::btree::lba_node_meta_t> : fmt::ostream_formatter {};
-template <> struct fmt::formatter<crimson::os::seastore::lba_manager::btree::lba_map_val_t> : fmt::ostream_formatter {};
 template <> struct fmt::formatter<crimson::os::seastore::lba_manager::btree::LBAInternalNode> : fmt::ostream_formatter {};
 template <> struct fmt::formatter<crimson::os::seastore::lba_manager::btree::LBALeafNode> : fmt::ostream_formatter {};
 #endif
